# Various performance improvements, especially for Range period and installations with many sites