* In the DataTableManipulator we requested the metadata for `all` websites. If an installation has many websites we ended up building the report metadata for all sites. This takes a long time, especially since some reports get site information eg `getFor('sitesearch')`. If an installation has 100 Piwik websites it resulted in 100 queries to `piwik_site` which resulted in 100 posted events in `Piwik::setSite` etc. 
* Post much less events. Eg we posted two events for each row in a GetSearchEngines datatable (which is slow as it will always notify all plugins). Instead we do now cache SearchEngineUrls / Names and SocialUrls. This saves about 100-200ms on a simple `Referrers.getSearchEngine` request if not more.
* When requesting a date range eg `date=last30&period=range` about 40% of the request time was spent in generating the sub periods. In one simple request we did this eg. 20 times for the same date range which took about 700ms on my server. As we cache it now it will take only about 35ms making all calls to range dates faster. I just this is more or less "only" a performance boost for `flat=1` as it looks like it generates the subperiods for each subtable. Note: If there was any long/endless running job this cache could result in a memory issue.

On our demo test server this call does now take only about 0.75s instead of 1.7s:
`index.php?date=2015-01-04,2015-02-02&module=Referrers&action=getSearchEngines&widget=1&isFooterExpandedInDashboard=true&flat=1&idSite=1&period=range`